backend/requirements/base.txt
Django>=5.1,<5.2
python-dotenv>=1.0,<2.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-allauth>=65.0,<66.0
django-money>=3.5,<4.0
django-filter>=24.0,<25.0
django-cors-headers>=4.6,<5.0
drf-spectacular>=0.28,<1.0
djangorestframework-camel-case>=1.4,<2.0
django-unfold>=0.40,<1.0
django-q2>=1.7,<2.0
redis>=5.2,<6.0
stripe>=11.0,<12.0
psycopg2-binary>=2.9,<3.0
Pillow>=11.0,<12.0
django-storages>=1.14,<2.0
boto3>=1.35,<2.0
py-moneyed>=3.0,<4.0
twilio>=9.0,<10.0
